requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.19
pytz==2024.7
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2
pandas==2.2.2
matplotlib==3.8.2
python-dotenv==1.0.0
